feat: implement zoomable image viewer to FileViewer
